arreglando cositas
