+++ app/templates/components/each-action-items.hbs
@@ -1,0 +1,3 @@
+{{#each actionItems as |unAction-item|}}
+{{edit-action-item unAction-item=unAction-item usuarios=usuarios actionItems=actionItems}}
+{{/each}}
