+++ app/templates/components/edit-action-item.hbs
@@ -1,0 +1,19 @@
+<div class="card-panel">
+  <a class="btn-floating waves-effect waves-light red right"  {{action 'borrarActionItem' unAction-item }}><i class="material-icons ">close</i></a>
+
+  <div >
+    {{tenpines/input-text label="descripción" value=unAction-item.descripcion}}
+  </div>
+  <div>
+    {{#power-select-multiple
+      options=usuariosSeleccionables
+      selected=unAction-item.responsables
+      placeholder="Seleccione los responsables del action item"
+      noMatchesMessage="No more pinos"
+      onchange=(action (mut unAction-item.responsables))
+    as |usuario|
+    }}
+      {{usuario.login}}
+    {{/power-select-multiple}}
+  </div>
+</div>
